% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuery.R
\name{CreateQuery}
\alias{CreateQuery}
\title{Create a query vector.}
\usage{
CreateQuery(qtext, terms, stemming = TRUE, language = "german", FUN = NA,
  max.distance = 0, ...)
}
\arguments{
\item{qtext, }{query-words (keywords) as string or vector with strings}

\item{terms, }{terms from a term-document matrix as vector with strings}

\item{stemming, }{determines if stemming should be applied to given keywords}

\item{language, }{language used for text stemming}

\item{FUN, }{function to apply weights to each term depending on the keywords}

\item{max.distance, }{maximum Levenshtein distance between a keyword and term}

\item{..., }{other parameters inquired as by the function}
}
\description{
This function creates a vector with weights from the given querywords and
term-document matrix.
}
\examples{
createQuery("term", c("vector", "with", "terms"))
createQuery("term", c("vector", "with", "terms"), max.distance=1, partial=F)
createQuery("term", c("vector", "with", "terms"),
            FUN=function(x, y) rep(1, length(y)))
}
\keyword{create_query}
\keyword{term_document_matrix}
\keyword{text_mining}
