% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVSM.R
\name{CreateVSM}
\alias{CreateVSM}
\title{Create a vector space model.}
\usage{
CreateVSM(text, ...)
}
\arguments{
\item{text, }{string or vector with strings}

\item{..., }{other parameters inquired as list "control" by the function
tm::TermDocumentMatrix()}
}
\description{
This function creates a term document matrix from the given string vector.
}
\examples{
CreateVSM(c("first sentence", "second sentence", "one more"))
}
\keyword{term_document_matrix}
\keyword{text_mining}
\keyword{vector_space_model}
