% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDocs.R
\name{SplitDocs}
\alias{SplitDocs}
\title{Split text into seperate blocks.}
\usage{
SplitDocs(documents, keep = ".?!:;_-", split = "[.:;!?]")
}
\arguments{
\item{documents, }{string or vector with strings to split}

\item{keep, }{defines special characters to be kept before splitting}

\item{split, }{defines characters used to split the given string(s)}
}
\description{
This function allows you to seperate a string or a vector with strings into 
single sentences.
}
\examples{
SplitDocs("Text: to be splitted into <single> blocks!!!")
}
\keyword{split}
\keyword{text}
